+++ Categories/Enchants.lua
@@ -64,5 +64,20 @@ table.insert(ns.CATEGORIES, {
         { id = 244009, rank = 2, qty = 1, enabled = true }, -- Enchant Boots - Farstrider's Hunt R2
         { id = 243952, rank = 1, qty = 1, enabled = true }, -- Enchant Boots - Lynx's Dexterity R1
         { id = 243953, rank = 2, qty = 1, enabled = true }, -- Enchant Boots - Lynx's Dexterity R2
+
+        -- Weapon
+        { header = "Weapon" },
+        { id = 243970, rank = 1, qty = 1, enabled = true }, -- Enchant Weapon - Jan'alai's Precision R1
+        { id = 243971, rank = 2, qty = 1, enabled = true }, -- Enchant Weapon - Jan'alai's Precision R2
+        { id = 243972, rank = 1, qty = 1, enabled = true }, -- Enchant Weapon - Berserker's Rage R1
+        { id = 243973, rank = 2, qty = 1, enabled = true }, -- Enchant Weapon - Berserker's Rage R2
+        { id = 243998, rank = 1, qty = 1, enabled = true }, -- Enchant Weapon - Worldsoul Aegis R1
+        { id = 243999, rank = 2, qty = 1, enabled = true }, -- Enchant Weapon - Worldsoul Aegis R2
+        { id = 244026, rank = 1, qty = 1, enabled = true }, -- Enchant Weapon - Flames of the Sin'dorei R1
+        { id = 244027, rank = 2, qty = 1, enabled = true }, -- Enchant Weapon - Flames of the Sin'dorei R2
+        { id = 244028, rank = 1, qty = 1, enabled = true }, -- Enchant Weapon - Acuity of the Ren'dorei R1
+        { id = 244029, rank = 2, qty = 1, enabled = true }, -- Enchant Weapon - Acuity of the Ren'dorei R2
+        { id = 244030, rank = 1, qty = 1, enabled = true }, -- Enchant Weapon - Arcane Mastery R1
+        { id = 244031, rank = 2, qty = 1, enabled = true }, -- Enchant Weapon - Arcane Mastery R2
     }
 })